Added UIMerchant, Update SceneObjectUID and WorldSceneDetailsCache
Added UIMerchant. Purchasing of items is WIP.
Added Merchant Templates.
Added ToRotation to SceneTeleporterDetails.
Added SceneObjectUID assignment to WorldSceneDetailsCache building. This will assign a unique ID to objects in World Scenes if they have the SceneObjectUID component. This is used primarily for Interactables.
Added Teleporter interactable type. You can assign a Target transform for player teleportation within the scene. (Scene teleportation is WIP)
Added more functionality to UITooltipButtons. These buttons will display a tooltip when you hover your mouse over them.

Removed Docker.cs

Fixed UI scaling issues.
Fixed CharacterInventorySystem character and inventory validation.
Fixed issue with the TargetController not generating Labels and/or Outlines properly for targets.

+++ FishMMO-Unity/Assets/Scripts/Shared/Network/Interactable/InteractableBroadcasts.cs
@@ -7,12 +7,12 @@ namespace FishMMO.Shared
 	/// </summary>
 	public struct InteractableBroadcast : IBroadcast
 	{
-		public long InteractableID;
+		public int InteractableID;
 	}
 
 	public struct AbilityCraftBroadcast : IBroadcast
 	{
-		public long InteractableID;
+		public int InteractableID;
 	}
 
 	public struct MerchantBroadcast : IBroadcast
